+++ src/constants.ts
@@ -114,6 +114,11 @@ export enum OSDOrientation {
 	AUTO = 2,
 }
 
+export enum OSDSignalInformation {
+	OFF = 1,
+	ON = 2,
+}
+
 export enum GSensorReading {
 	LANDSCAPE = 1, // NEC logo at the bottom
 	WRONG_PORTRAIT = 2, // NEC logo on the left, not supported