+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- OSD signal information
+
 ## [0.4.1] - 2021-01-26
 ### Fixed
 - Corrected CM3 max schedule index to 29